% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_char_vars.R
\name{compare_char_vars}
\alias{compare_char_vars}
\title{Compare a set of repro character values to original character values to check for matches}
\usage{
compare_char_vars(
  df_orig,
  df_repro,
  checkvars,
  name_orig = "original",
  name_repro = "repro",
  name_diff = "values_identical"
)
}
\arguments{
\item{df_orig}{The data frame containing the original character values}

\item{df_repro}{The data frame containing the reproduced character values}

\item{checkvars}{A character vector containing variable names to compare}

\item{name_orig}{A character string containing the column name for the original value. Default \code{"original"}.}

\item{name_repro}{A character string containing the column name for the reproduced value. Default \code{"repro"}.}

\item{name_diff}{A character string containing the column name for the string comparison. Default \code{"values_identical"}.}
}
\value{
A data frame containing five named variables: \code{compared}, the value of \code{name_orig}, the value of \code{name_repro}, the value of \code{name_diff}, and \code{note}. The values in the \code{compared} column will be the values of \code{checkvars}.
}
\description{
This function takes two data frames, an original result and a reproduced result, and creates a data frame containing the results of string matches of a set of variables.
This function uses the base R function \code{identical} to compare character strings.
}
