% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_vars.R
\name{compare_vars}
\alias{compare_vars}
\title{Compare a set of repro values to original values using pcterr()}
\usage{
compare_vars(
  df_orig,
  df_repro,
  checkvars,
  name_orig = "original",
  name_repro = "repro",
  name_diff = "percent_error"
)
}
\arguments{
\item{df_orig}{The data frame containing the original values}

\item{df_repro}{The data frame containing the reproduced values}

\item{checkvars}{A character vector containing variable names to compare}

\item{name_orig}{A character string containing the column name for the original value. Default \code{"original"}.}

\item{name_repro}{A character string containing the column name for the reproduced value. Default \code{"repro"}.}

\item{name_diff}{A character string containing the column name for the percent error. Default \code{"percent_error"}.}
}
\value{
A data frame containing five named variables: \code{compared}, the value of \code{name_orig}, the value of \code{name_repro}, the value of \code{name_diff}, and \code{note}. The values in the \code{compared} column will be the values of \code{checkvars}.
}
\description{
This function takes two data frames, an original result and a reproduced result, and creates a data frame containing the percent error of a set of variables.
This function makes use of the \code{pcterr} function.
}
