% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcterr.R
\name{pcterr}
\alias{pcterr}
\title{Calculate the percent error between two values}
\usage{
pcterr(origval, repval)
}
\arguments{
\item{origval}{The original numeric value to compare against}

\item{repval}{The recalculated or reproduced numeric value to be compared}
}
\value{
A list containing the percent error (\code{pcterror}) and an informative text description (\code{note})
}
\description{
This function calculates the relative error between two values, where one value is deemed to be the original value, and the other is the recalculated/reproduced value. See definition on \href{https://en.wikipedia.org/wiki/Approximation_error#Formal_Definition}{Wikipedia} for reference.
The relative error is expressed as a percentage of the original value.
The function returns a percent error of NA if the original value is 0.
The function returns a percent error of NA if either the original value or recalculated value is \code{NA}, \code{NULL}, \code{NaN}, or non-numeric.
}
\examples{
pcterr(1.0, 1.0)

pcterr(1.0, 1.09)

pcterr(1.0, 1.10)
}
